EasyStream.jl

EasyStream.distributionByClass

EasyStream.prequentialAnalyze

EasyStream.distributionByClass — Method

distributionByClass(xg, yg, pred, predicted_y; colors = [:pink, :lightblue])

Generates a graph plotting the samples and shows the spatial limit for what the model considers contained in each class.

X is the variable contains the points to be ploted in the scatter and predicted_y determine the classes of these points.

Variables xg and xy determine the grid of the space. More points contained in these variables mean a more detailed space.

Variable pred contains the model that will be used to determined the spacial classification.

EasyStream.prequentialAnalyze — Method

prequentialAnalyze(ŷ, y, steps, meansure)

Generates a graph plotting the results of the meansure on intervals of time defined to a determined number of smaples length(ŷ) / steps.

ŷ and y are the values wich will be used in the meansure function to be result the plot points.

steps defines how much grainy will be the graph.